+++ packages/webcodecs/src/can-reencode-video-track.ts
@@ -1,5 +1,5 @@
 import type {VideoTrack} from '@remotion/media-parser';
-import type {ConvertMediaVideoCodec} from './codec-id';
+import type {ConvertMediaVideoCodec} from './get-available-video-codecs';
 import {getVideoDecoderConfigWithHardwareAcceleration} from './video-decoder-config';
 import {getVideoEncoderConfig} from './video-encoder-config';
