+++ packages/webcodecs/src/get-available-video-codecs.ts
@@ -21,25 +21,3 @@ export const getAvailableVideoCodecs = (
 
 	throw new Error(`Unsupported container: ${container satisfies never}`);
 };
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-const availableAudioCodecs = ['opus', 'aac', 'wav'] as const;
-export const getAvailableAudioCodecs = (
-	container: ConvertMediaContainer,
-): ConvertMediaAudioCodec[] => {
-	if (container === 'mp4') {
-		return ['aac'];
-	}
-
-	if (container === 'webm') {
-		return ['opus'];
-	}
-
-	if (container === 'wav') {
-		return ['wav'];
-	}
-
-	throw new Error(`Unsupported container: ${container satisfies never}`);
-};
-
-export type ConvertMediaAudioCodec = (typeof availableAudioCodecs)[number];
