+++ packages/webcodecs/src/get-default-audio-codec.ts
@@ -1,4 +1,4 @@
-import type {ConvertMediaAudioCodec} from './codec-id';
+import type {ConvertMediaAudioCodec} from './get-available-audio-codecs';
 import type {ConvertMediaContainer} from './get-available-containers';
 
 export const getDefaultAudioCodec = ({
