+++ packages/webcodecs/src/on-audio-track-handler.ts
@@ -1,5 +1,5 @@
 import type {AudioTrack, LogLevel} from '@remotion/media-parser';
-import type {ConvertMediaAudioCodec} from './codec-id';
+import type {ConvertMediaAudioCodec} from './get-available-audio-codecs';
 import type {ConvertMediaContainer} from './get-available-containers';
 
 export type AudioOperation =
