+++ packages/webcodecs/src/on-video-track-handler.ts
@@ -1,6 +1,6 @@
 import type {LogLevel, VideoTrack} from '@remotion/media-parser';
-import type {ConvertMediaVideoCodec} from './codec-id';
 import type {ConvertMediaContainer} from './get-available-containers';
+import type {ConvertMediaVideoCodec} from './get-available-video-codecs';
 
 export type VideoOperation =
 	| {type: 'reencode'; videoCodec: ConvertMediaVideoCodec}
